Added and implemented SupplierSuccess objects in the service and controller layers of affected methods.

+++ eindwerkJava2/src/main/java/com/example/eindwerkJava2/controller/SupplierController.java
@@ -5,10 +5,13 @@ import com.example.eindwerkJava2.repositories.SupplierRepository;
 import com.example.eindwerkJava2.service.CitiesService;
 import com.example.eindwerkJava2.service.CountriesService;
 import com.example.eindwerkJava2.service.SupplierService;
+import com.example.eindwerkJava2.wrappers.SuccessObject;
+import com.example.eindwerkJava2.wrappers.SupplierSuccess;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 @Controller
 public class SupplierController {
@@ -21,36 +24,48 @@ public class SupplierController {
     private CountriesService countriesService;
 
     @GetMapping("/supplier")
-    public String viewSuppliers(Model model){
-        model.addAttribute("suppliersList", supplierService.getAllSuppliers());
+    public String viewSuppliers(Model model) {
+        SupplierSuccess supplierSuccess = supplierService.getAllSuppliers();
+        if (supplierSuccess.getIsSuccessfull()) {
+            model.addAttribute("suppliersList", supplierSuccess.getSuppliers());
+        } else {
+            model.addAttribute("error", supplierSuccess.getMessage());
+        }
         return "supplier";
     }
 
     @GetMapping("/new/supplier")
-    public String showNewSupplierForm(Model model){
+    public String showNewSupplierForm(Model model) {
         Supplier supplier = new Supplier();
         model.addAttribute("supplier", new Supplier());
         model.addAttribute("citiesList", citiesService.getAllCities().getCities());
         model.addAttribute("countriesList", countriesService.getAllCountries());
         return "/forms/form_supplier";
     }
+
     @GetMapping("/generatePdf")
-    public String generatePdf(){
+    public String generatePdf() {
         Supplier supplier = new Supplier();
         supplierService.makePdf();
         return "redirect:/supplier";
     }
 
     @PostMapping("/saveSupplier")
-    public String saveSupplier(@ModelAttribute("supplier") Supplier supplier){
-        this.supplierService.saveSupplier(supplier);
-        return "redirect:/supplier";
+    public String saveSupplier(@ModelAttribute("supplier") Supplier supplier, RedirectAttributes redirAttr, Model model) {
+        SuccessObject isSaveSuccessful = this.supplierService.saveSupplier(supplier);
+        if (isSaveSuccessful.getIsSuccessfull()) {
+            redirAttr.addFlashAttribute("success", isSaveSuccessful.getMessage());
+            return "redirect:/supplier";
+        } else {
+            model.addAttribute("error", isSaveSuccessful.getMessage());
+            return "/forms/form_supplier";
+        }
     }
 
-
     @GetMapping("edit/supplier/{supplierId}")
-    public String showEditSupplierForm(@PathVariable("supplierId") Long supplierId, Model model){
-        Supplier supplier = supplierService.findById(supplierId).get();
+    public String showEditSupplierForm(@PathVariable("supplierId") Long supplierId, Model model) {
+        SupplierSuccess findSupplierSuccess = supplierService.findById(supplierId);
+        Supplier supplier = findSupplierSuccess.getSupplier();
         model.addAttribute("supplier", supplier);
         model.addAttribute("citiesList", citiesService.getAllCities().getCities());
         model.addAttribute("countriesList", countriesService.getAllCountries());
@@ -58,17 +73,21 @@ public class SupplierController {
     }
 
     @GetMapping("delete/supplier/{supplierId}")
-    public String deleteSupplier(@PathVariable("supplierId") Long supplierId, Model model){
-        Supplier supplier = supplierService.findById(supplierId).get();
-        this.supplierService.deleteSupplier(supplier);
+    public String deleteSupplier(@PathVariable("supplierId") Long supplierId, RedirectAttributes redirAttr) {
+        SupplierSuccess findSupplierSuccess = supplierService.findById(supplierId);
+        if (findSupplierSuccess.getIsSuccessfull()) {
+            Supplier supplier = findSupplierSuccess.getSupplier();
+            SuccessObject toBeDeletedSupplier = this.supplierService.deleteSupplier(supplier);
+            if (toBeDeletedSupplier.getIsSuccessfull()) {
+                redirAttr.addFlashAttribute("success", toBeDeletedSupplier.getMessage());
+            } else {
+                redirAttr.addFlashAttribute("error", toBeDeletedSupplier.getMessage());
+            }
+        } else {
+            redirAttr.addFlashAttribute("error", findSupplierSuccess.getMessage());
+        }
         return "redirect:/supplier";
     }
-
-
-
-
-
-
 
 
 }
